remove and rename old site tree components (#19478)
